Removed Dependencies header under Clone section
Removed Dependencies header to make it clear that the dependencies installation was a part of the Clone section and not that PyPI, Clone, and Dependencies were the three steps one had to take to get the code. Will add a dedicated dependencies section further down that will outline the dependencies used, why they were chosen, etc.